fe: fix cyling importing warning
