+++ app/client/mobile/constants/settings.ts
@@ -1,0 +1,19 @@
+export enum SETTING_KEY {
+  LANGUAGE = 'LANGUAGE',
+  DARK_MODE = 'DARK_MODE',
+  NOTIFICATION_COMMENT = 'NOTIFICATION_COMMENT',
+  NOTIFICATION_VOTE = 'NOTIFICATION_VOTE',
+  NOTIFICATION_FOLLOW = 'NOTIFICATION_FOLLOW',
+}
+
+export namespace SETTING_VALUE {
+  export enum LANGUAGE {
+    VIETNAMESE = 'VIETNAMESE',
+    ENGLISH = 'ENGLISH',
+  }
+
+  export enum BOOLEAN {
+    TRUE = 'TRUE',
+    FALSE = 'FALSE',
+  }
+}
