+++ app/client/mobile/utils/converter.ts
@@ -1,10 +1,9 @@
-import { SETTING_VALUE } from "@/slices/setting.slice";
+import { SETTING_VALUE } from "@/constants/settings";
 
 export const getBooleanValueFromSetting = (value: SETTING_VALUE.BOOLEAN) => {
   return value === SETTING_VALUE.BOOLEAN.TRUE ? true : false;
-} 
+};
 
 export const getSettingFromBooleanValue = (value: boolean) => {
   return value ? SETTING_VALUE.BOOLEAN.TRUE : SETTING_VALUE.BOOLEAN.FALSE;
-} 
-
+};
